+++ src/main/kotlin/io/openfuture/chain/crypto/signature/SignatureManager.kt
@@ -1,15 +1,34 @@
 package io.openfuture.chain.crypto.signature
 
+import io.openfuture.chain.crypto.domain.ECKey
+import io.openfuture.chain.crypto.util.HashUtils
 import org.springframework.stereotype.Component
+import java.util.*
 
 @Component
 class SignatureManager {
 
-    // will be removed when will be implemented by another developer
+    /**
+     * Sign data with ECDSA
+     *
+     * @param data Information to sign
+     * @param privateKey Private key of signer
+     * @return Base64 encoded signature
+     */
     fun sign(data: ByteArray, privateKey: ByteArray): String =
-        "21334324324235435454565656565656546356fsdafdasfdasfd324134324432"
+        Base64.getEncoder().encodeToString(ECKey(privateKey).sign(HashUtils.sha256(data)))
 
-    // will be removed when will be implemented by another developer
-    fun verify(data: ByteArray, signature: String, publicKey: ByteArray) = true
+    /**
+     * Verify ECDSA signature
+     *
+     * @param data Information that was signed
+     * @param signature Base64 encoded signature
+     * @return True if valid signature. False otherwise
+     */
+    fun verify(data: ByteArray, signature: String, publicKey: ByteArray): Boolean {
+        val ecKey = ECKey(publicKey, false)
+        val decodedSign = Base64.getDecoder().decode(signature)
+        return ecKey.verify(HashUtils.sha256(data), decodedSign)
+    }
 
 }